Update manage-payg-transition.ps1
adding a variable $SkipDownload default value of $False. 
this variable will be converted into a parameter with default value of $False. 
the idea of this variable/parameter is to assist the execution of the script where companies block downloads from URLs. In this case the customer will need to download the 2 .ps1 files manually and place them into a subfolder named "PayTransitionDownloads"